+++ index.js
@@ -5,11 +5,11 @@ const { db } = require('./src/db');
 db.migrate.latest({ directory: path.join(__dirname, '/migrations') }); // TODO: await this
 
 // Expose the modules
-exports.Admin = require('./src/modules/admin');
-exports.Chores = require('./src/modules/chores');
-exports.Hearts = require('./src/modules/hearts');
-exports.Polls = require('./src/modules/polls');
-exports.Things = require('./src/modules/things');
+exports.Admin = require('./src/core/admin');
+exports.Chores = require('./src/core/chores');
+exports.Hearts = require('./src/core/hearts');
+exports.Polls = require('./src/core/polls');
+exports.Things = require('./src/core/things');
 
 // Expose the config and utilities
 // TODO: make config more... configurable
