+++ src/core/admin.js
