+++ src/core/hearts.js
