+++ src/core/polls.js
