+++ src/core/power.js
